getting user metadata slowly ehehehe
