+++ src/views/chat/useUserMetadataForChannel.ts
@@ -1,3 +1,4 @@
+import { handleEvent, useRelayPool } from 'app/lib/nostr'
 import { getNostrEvent } from 'app/lib/nostr/getNostrEvent'
 import { useEffect } from 'react'
 
@@ -9,20 +10,32 @@ export const useUserMetadataForChannel = (channelId: string) => {
   //   const userMetadata = useStore((s) => s.userMetadata)
   const messages = useMessagesForChannel(channelId)
 
+  const relaypool = useRelayPool()
   useEffect(() => {
     // Extract the list of unique public keys of the senders of the messages
     const uniquePubkeys = [
       ...new Set(messages.map((message) => message.pubkey)),
     ] as string[]
 
-    getNostrEvent([{ kinds: [0], authors: uniquePubkeys }]).then((stream) => {
-      console.log('maybe got something', stream)
-    })
+    if (uniquePubkeys.length === 0) {
+      console.log('no unique pubkeys. returning')
+    }
 
     // Now fetch metadata for each pubkey
-    // uniquePubkeys.forEach((pubkey) => {
-    //   if (!userMetadata.has(pubkey)) {
-    //   }
-    // })
-  }, [messages])
+    relaypool.relayPool?.subscribe(
+      [
+        {
+          kinds: [0],
+          authors: uniquePubkeys,
+        },
+      ],
+      relaypool.relays.map((relay) => relay.url),
+      (event) => {
+        handleEvent(event)
+      },
+      (eose) => {
+        console.log('eose for ', channelId)
+      }
+    )
+  }, [messages, channelId])
 }
